+++ test/vaelii/special_table_test.clj
@@ -111,6 +111,8 @@
     [disjoint_metatype                 [:integrate :disintegrate :rebuild :wff]       nil                        true]
     [sibling_disjoint                  [:integrate :disintegrate :rebuild :wff]       nil                        true]
     [siblingDisjointException         [:integrate :disintegrate :rebuild :wff]       nil                        true]
+    [covering                         [:integrate :disintegrate :rebuild :wff]       nil                        true]
+    [partitionedInto                  [:integrate :disintegrate :rebuild :wff]       nil                        true]
     [transitive                       [:integrate :disintegrate :rebuild :wff]       :transitive                true]
     [symmetric                        [:integrate :disintegrate :rebuild :wff]       :symmetric                 true]
     [commutative                      [:integrate :disintegrate :rebuild :wff]       :commutative               true]